requirements.txt
fastapi
load_dotenv